New data model to optimize querying performance in multi-tenancy
