+++ pkgs/backend/src/entities/wms/stock_entries.rs
@@ -7,8 +7,9 @@ use sea_orm::entity::prelude::*;
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
-    pub tenant_id: Option<i32>,
+    pub tenant_id: i32,
     pub lot_id: i32,
+    pub stock_id: i32,
     pub quantity: i32,
     pub status: i32,
     pub expired_at: Option<DateTimeUtc>,
